+++ src/app/_models/user.model.ts
@@ -1,0 +1,10 @@
+export class User {
+    username: string;
+    location: string;
+    email: string;
+    website: string;
+    birthday: string;
+    image: string;
+    roles: string;
+    constructor() {}
+}
